+++ src/application/model/Vand.java
@@ -1,7 +1,7 @@
 package application.model;
 
 public class Vand extends Råvarer {
-    public Vand(String navn, String lokation, int mængde) {
+    public Vand(String navn, String lokation, double mængde) {
         super(navn, lokation, mængde);
     }
